change to Composition from Inheritance
Since we're not really extending the TOO class (we're really using
it without changing it's behaviour), it make more sense to compose
the object (i.e. instanciate the TOO class as a property of the
SwiftAPI class. This makes even more sense when we do the same thing
with the TOORequests class.